requirement.txt
streamlit
langchain-community
qdrant-client
langchain-huggingface
python-dotenv
sentence-transformers
google-genai